Use Netlify API to wait for deploy previews
The previous wait-for-netlify-action was waiting for a 200 response on the
deploy preview URL, but this will always succeed after the initial build. Any
subsequent change to the pull request will see the Lighthouse report built on a
stale site and therefore the Lighthouse report will be stale too.

The patch migrates to a different action, andrewnicols/wait-for-netlify-preview,
which uses the Netlify API locate the Deploy ID for the most recent build of a
branch, and will wait for the build to complete or fail.

This does require an API key from Netlify to function.